+++ src/main/java/io/github/xf8b/morefeatures/items/armor/lapis/LapisChestplate.java
@@ -3,16 +3,13 @@ package io.github.xf8b.morefeatures.items.armor.lapis;
 import io.github.xf8b.morefeatures.MoreFeatures;
 import io.github.xf8b.morefeatures.MoreFeaturesArmorMaterial;
 import io.github.xf8b.morefeatures.config.MoreFeaturesConfig;
+import io.github.xf8b.morefeatures.util.handler.TickHandler;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.EquipmentSlotType;
 import net.minecraft.item.ArmorItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import net.minecraftforge.event.TickEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
 
-@Mod.EventBusSubscriber(modid = MoreFeatures.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class LapisChestplate extends ArmorItem {
     public LapisChestplate() {
         super(MoreFeaturesArmorMaterial.LAPIS, EquipmentSlotType.CHEST, new Properties()
@@ -23,20 +20,8 @@ public class LapisChestplate extends ArmorItem {
 
     @Override
     public void onArmorTick(ItemStack stack, World world, PlayerEntity player) {
-        if (count == 2400) {
+        if (TickHandler.clientTicksPassed % 2400 == 0) {
             player.giveExperiencePoints(MoreFeaturesConfig.lapisArmorExperienceGiven);
-        }
-    }
-
-    public static int count = 0;
-
-    @SubscribeEvent
-    public static void tickCounter(TickEvent.ClientTickEvent event) {
-        if (event.phase == TickEvent.Phase.END) {
-            count++;
-            if (count > 2400) {
-                count = 0;
-            }
         }
     }
 }