+++ src/main/java/io/github/xf8b/morefeatures/items/tools/corrupted/CorruptedHoe.java
@@ -40,7 +40,7 @@ public class CorruptedHoe extends HoeItem {
             tooltip.add(new StringTextComponent("The attack speed will be random."));
             tooltip.add(new StringTextComponent("Good luck, and pray to the RNG gods."));
         } else {
-            tooltip.add(new StringTextComponent("Hold " + "\u00A7e" + "SHIFT " + "\u00A77" + "to see info."));
+            tooltip.add(new StringTextComponent("Hold " + "\u00A7e" + "SHIFT " + "\u00A7r" + "to see info."));
         }
         super.addInformation(stack, worldIn, tooltip, flagIn);
     }